requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
csh==0.1.2
gunicorn==19.1.1
itsdangerous==0.24
meld3==1.0.0
python-dateutil==2.2
python-ldap==2.4.15
requests==2.3.0
sasl==0.1.3
six==1.7.3
supervisor==3.1.3
wsgiref==0.1.2
